+++ test/locale/ta.js
@@ -313,12 +313,13 @@ exports['locale:ta'] = {
     },
 
     'meridiem' : function (test) {
+        test.equal(moment([2011, 2, 23,  0, 30]).format('a'), ' யாமம்', '(after) midnight');
         test.equal(moment([2011, 2, 23,  2, 30]).format('a'), ' வைகறை', 'before dawn');
         test.equal(moment([2011, 2, 23,  9, 30]).format('a'), ' காலை', 'morning');
-        test.equal(moment([2011, 2, 23, 14, 30]).format('a'), ' நண்பகல்', 'during day');
+        test.equal(moment([2011, 2, 23, 14, 30]).format('a'), ' எற்பாடு', 'during day');
         test.equal(moment([2011, 2, 23, 17, 30]).format('a'), ' எற்பாடு', 'evening');
         test.equal(moment([2011, 2, 23, 19, 30]).format('a'), ' மாலை', 'late evening');
-        test.equal(moment([2011, 2, 23, 21, 20]).format('a'), ' இரவு', 'night');
+        test.equal(moment([2011, 2, 23, 23, 30]).format('a'), ' யாமம்', '(before) midnight');
         test.done();
     },
 
@@ -348,6 +349,20 @@ exports['locale:ta'] = {
             test.equal(testMoment.date(), i,
                     'lenient ordinal parsing of number ' + i + ' date check');
         }
+        test.done();
+    },
+
+    'meridiem invariant' : function (test) {
+        var h, m, t1, t2;
+        for (h = 0; h < 24; ++h) {
+            for (m = 0; m < 60; m += 15) {
+                t1 = moment.utc([2000, 0, 1, h, m]);
+                t2 = moment(t1.format('A h:mm'), 'A h:mm');
+                test.equal(t2.format('HH:mm'), t1.format('HH:mm'),
+                        'meridiem at ' + t1.format('HH:mm'));
+            }
+        }
+
         test.done();
     },